+++ Plugin.BluetoothLE/Platforms/Android/Internals/InternalScanRecord.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Collections.Generic;
 using System.Text;
 
+
 namespace Plugin.BluetoothLE.Internals
 {
     public class InternalScanRecord
@@ -10,6 +11,7 @@ namespace Plugin.BluetoothLE.Internals
         public static InternalScanRecord Parse(byte[] scanRecord)
         {
             var sr = new InternalScanRecord();
+            var mfData = new List<ManufacturerData>();
             var index = 0;
             var others = new List<AdRecord>();
 
@@ -48,7 +50,11 @@ namespace Plugin.BluetoothLE.Internals
                         break;
 
                     case AdvertisementRecordType.ManufacturerSpecificData:
-                        sr.ManufacturerData = data;
+                        var manufacturerId = ((data[1] & 0xFF) << 8) + (data[0] & 0xFF);
+                        var manufacturerDataBytes = new byte[data.Length - 2];
+                        Array.Copy(data, 2, manufacturerDataBytes, 0, data.Length-2);
+
+                        mfData.Add(new ManufacturerData((ushort) manufacturerId, manufacturerDataBytes));
                         break;
 
                     default:
@@ -64,12 +70,14 @@ namespace Plugin.BluetoothLE.Internals
                 .ToList()
                 .ForEach(sr.ServiceUuids.Add);
 
+            sr.ManufacturerData = mfData.ToArray();
+
             return sr;
         }
 
 
         public string LocalName { get; private set; }
-        public byte[] ManufacturerData { get; private set; }
+        public ManufacturerData[] ManufacturerData { get; private set; }
         public bool IsConnectable { get; private set; }
         public int TxPower { get; private set; }
         public IList<Guid> ServiceUuids { get; } = new List<Guid>();