+++ Plugin.BluetoothLE/Platforms/Uwp/AdvertisementData.cs
@@ -10,7 +10,7 @@ namespace Plugin.BluetoothLE
     {
         readonly BluetoothLEAdvertisementReceivedEventArgs adData;
         readonly Lazy<Guid[]> serviceUuids;
-        readonly Lazy<byte[]> manufacturerData;
+        readonly Lazy<ManufacturerData[]> manufacturerData;
         readonly Lazy<int> txPower;
 
 
@@ -18,7 +18,7 @@ namespace Plugin.BluetoothLE
         {
             this.adData = args;
 
-            this.manufacturerData = new Lazy<byte[]>(() => args.Advertisement.GetManufacturerSpecificData());
+            this.manufacturerData = new Lazy<ManufacturerData[]>(() => args.Advertisement.GetManufacturerSpecificData());
             this.serviceUuids = new Lazy<Guid[]>(() => args.Advertisement.ServiceUuids.ToArray());
             this.txPower = new Lazy<int>(() => args.Advertisement.GetTxPower());
         }
@@ -31,7 +31,7 @@ namespace Plugin.BluetoothLE
                                      this.adData.AdvertisementType == BluetoothLEAdvertisementType.ConnectableUndirected;
 
         public IReadOnlyList<byte[]> ServiceData { get; } = null;
-        public byte[] ManufacturerData => this.manufacturerData.Value;
+        public ManufacturerData[] ManufacturerData => this.manufacturerData.Value;
         public Guid[] ServiceUuids => this.serviceUuids.Value;
         public int TxPower => this.txPower.Value;
     }